requirements.txt
appdirs==1.4.4
asgiref==3.4.1
astroid==2.6.2
backcall==0.2.0
black==21.6b0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
decorator==5.0.9
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.66.0
fastapi-auth0==0.2.1
h11==0.12.0
idna==2.10
ipython==7.25.0
ipython-genutils==0.2.0
isort==5.9.1
jedi==0.18.0
lazy-object-proxy==1.6.0
matplotlib-inline==0.1.2
mccabe==0.6.1
mypy-extensions==0.4.3
parso==0.8.2
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pyasn1==0.4.8
pydantic==1.8.2
Pygments==2.9.0
pylint==2.9.3
python-dotenv==0.18.0
python-jose==3.3.0
regex==2021.7.6
requests==2.25.1
rsa==4.7.2
six==1.16.0
starlette==0.14.2
toml==0.10.2
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
wcwidth==0.2.5
wrapt==1.12.1
